requirements.txt
google-cloud-texttospeech==2.16.3
google-cloud-storage==2.10.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
librosa==0.10.1
soundfile==0.12.1
numpy==1.24.3
python-dotenv==1.0.0
aiofiles==23.2.0
